# src/main/resources/static/openapi/ignitr.yaml
openapi: 3.0.3
info:
  title: Ignitr API
  version: 1.0.0

tags:
  - name: Health
    description: System health checks
  - name: Sparks
    description: Operations for managing Sparks
  - name: Reasons
    description: Operations for managing Reasons attached to Sparks

paths:
  /api/health:
    get:
      tags:
        - Health
      summary: Health check
      description: Returns OK if the service is running.
      responses:
        "200":
          description: Service is healthy.
          content:
            text/plain:
              schema:
                type: string
                example: OK

  /api/sparks:

    get:
      tags:
        - Sparks
      summary: Search and filter Sparks
      description: |
        Retrieves Sparks using optional filters. Supports pagination.
        - If **title** is provided: returns Sparks whose titles contain the value (case-insensitive).
        - If **parentId** is provided: returns Sparks that are children of that parent.
        - If **parentId** value is "ROOT": returns only top-level Sparks (no parent).
        - If both are omitted: returns all Sparks.

      parameters:
        - name: title
          in: query
          required: false
          schema:
            type: string
          description: Case-insensitive partial match on title.
        - name: parentId
          in: query
          required: false
          schema:
            type: string
          description: Return only Sparks that have this parent id. Use special value "ROOT" for top-level Sparks.
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Page index (0-based).
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Page size (1–100).
      responses:
        "200":
          description: Paginated list of Sparks.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SparkPage'
        "400":
          description: Validation error. Returns VALIDATION_ERROR (for invalid pagination parameters).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      tags:
        - Sparks
      summary: Create a top-level Spark
      description: Creates a new Spark with no parent (top-level).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSparkRequest'
      responses:
        "201":
          description: Spark created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Spark'
        "400":
          description: Validation error. Returns VALIDATION_ERROR.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        "409":
          description: Spark with same title already exists. Returns SPARK_ALREADY_EXISTS.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

  /api/sparks/{id}:
    get:
      tags:
        - Sparks
      summary: Get a Spark by id
      description: Returns a single Spark by its id.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID of the Spark.
      responses:
        "200":
          description: Spark found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Spark'
        "404":
          description: Spark not found. Returns SPARK_NOT_FOUND.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    put:
      tags:
        - Sparks
      summary: Update an existing Spark
      description: Updates the title and description of an existing Spark.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID of the Spark to update.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSparkRequest'
      responses:
        '200':
          description: Spark updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Spark'
        '400':
          description: Invalid input. Returns VALIDATION_ERROR.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Spark not found. Returns SPARK_NOT_FOUND.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Spark title must be unique. Returns SPARK_ALREADY_EXISTS.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    patch:
      tags:
        - Sparks
      summary: Partially update an existing Spark
      description: Updates one or more fields of an existing Spark. Only provided fields are changed.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID of the Spark to patch.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchSparkRequest'
      responses:
        '200':
          description: Spark patched successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Spark'
        '400':
          description: Invalid input. Returns VALIDATION_ERROR.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Spark not found. Returns SPARK_NOT_FOUND.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Spark title must be unique. Returns SPARK_ALREADY_EXISTS.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    delete:
      tags:
        - Sparks
      summary: Delete a Spark
      description: Deletes the Spark with the given id according to the selected delete mode.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID of the Spark to delete.
        - name: mode
          in: query
          required: false
          description: Delete mode. Defaults to CASCADE if not specified.
          schema:
            type: string
            enum:
              - CASCADE
              - PROMOTE
            default: CASCADE
      responses:
        "204":
          description: Spark deleted successfully. No content is returned.
        "400":
          description: Validation error. Returns VALIDATION_ERROR.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        "404":
          description: Spark not found. Returns SPARK_NOT_FOUND.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

  /api/sparks/{parentId}/children:
    get:
      tags:
        - Sparks
      summary: Get direct children of a Spark
      description: Returns all Sparks whose parent is the given Spark id.
      parameters:
        - name: parentId
          in: path
          required: true
          schema:
            type: string
          description: ID of the parent Spark.
      responses:
        "200":
          description: Children found (may be an empty list if none).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Spark'
        "404":
          description: Parent Spark not found. Returns SPARK_NOT_FOUND.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      tags:
        - Sparks
      summary: Create a child Spark
      description: Creates a Spark as a child of the given Spark.
      parameters:
        - name: parentId
          in: path
          required: true
          schema:
            type: string
          description: ID of the parent Spark.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSparkRequest'
      responses:
        "201":
          description: Child Spark created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Spark'
        "400":
          description: Validation error. Returns VALIDATION_ERROR.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        "404":
          description: Parent Spark not found. Returns SPARK_NOT_FOUND.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        "409":
          description: Spark with same title already exists. Returns SPARK_ALREADY_EXISTS.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

  /api/sparks/{id}/tree:
    get:
      tags:
        - Sparks
      summary: Get spark tree
      description: Returns the Spark with the given id and all its descendants as a recursive tree structure.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID of the root Spark.
      responses:
        '200':
          description: Spark tree returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SparkTree'
        '404':
          description: Spark not found. Returns SPARK_NOT_FOUND.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

  /api/sparks/{sparkId}/reasons:
    get:
      tags:
        - Reasons
      summary: List Reasons for a Spark
      description: Returns a paginated list of Reasons attached to the specified Spark. Optionally filters by reason type.
      parameters:
        - name: sparkId
          in: path
          required: true
          schema:
            type: string
          description: ID of the Spark whose Reasons should be returned.
        - name: type
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ReasonType'
          description: Filter results by Reason type (case-insensitive).
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Page index (0-based).
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
          description: Page size (1–100).
      responses:
        '200':
          description: Reasons returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonPage'
        '400':
          description: Validation error. Returns VALIDATION_ERROR.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Spark not found. Returns SPARK_NOT_FOUND.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      tags:
        - Reasons
      summary: Create a Reason for a Spark
      description: Creates a new Reason attached to the specified Spark.
      parameters:
        - name: sparkId
          in: path
          required: true
          schema:
            type: string
          description: ID of the Spark receiving the new reason.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReasonRequest'
      responses:
        '201':
          description: Reason created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reason'
        '400':
          description: Validation error. Returns VALIDATION_ERROR.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Spark not found. Returns SPARK_NOT_FOUND.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    delete:
      tags:
        - Reasons
      summary: Delete all Reasons for a Spark
      description: Removes every Reason attached to the specified Spark.
      parameters:
        - name: sparkId
          in: path
          required: true
          schema:
            type: string
          description: ID of the Spark whose Reasons will be removed.
      responses:
        '204':
          description: Reasons deleted successfully. No content is returned.
        '404':
          description: Spark not found. Returns SPARK_NOT_FOUND.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

  /api/sparks/{sparkId}/reasons/{reasonId}:
    get:
      tags:
        - Reasons
      summary: Get a Reason by id
      description: Returns a single Reason belonging to the specified Spark.
      parameters:
        - name: sparkId
          in: path
          required: true
          schema:
            type: string
          description: ID of the Spark that owns the reason.
        - name: reasonId
          in: path
          required: true
          schema:
            type: string
          description: ID of the Reason to retrieve.
      responses:
        '200':
          description: Reason found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reason'
        '404':
          description: Reason or Spark not found. Returns REASON_NOT_FOUND or SPARK_NOT_FOUND.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    put:
      tags:
        - Reasons
      summary: Update a Reason
      description: Replaces the content and type of an existing Reason.
      parameters:
        - name: sparkId
          in: path
          required: true
          schema:
            type: string
          description: ID of the Spark that owns the Reason.
        - name: reasonId
          in: path
          required: true
          schema:
            type: string
          description: ID of the Reason to update.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReasonRequest'
      responses:
        '200':
          description: Reason updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reason'
        '400':
          description: Validation error. Returns VALIDATION_ERROR.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Reason or Spark not found. Returns REASON_NOT_FOUND or SPARK_NOT_FOUND.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    delete:
      tags:
        - Reasons
      summary: Delete a Reason
      description: Deletes a single reason belonging to the specified Spark.
      parameters:
        - name: sparkId
          in: path
          required: true
          schema:
            type: string
          description: ID of the Spark that owns the Reason.
        - name: reasonId
          in: path
          required: true
          schema:
            type: string
          description: ID of the Reason to delete.
      responses:
        '204':
          description: Reason deleted successfully. No content is returned.
        '404':
          description: Reason or Spark not found. Returns REASON_NOT_FOUND or SPARK_NOT_FOUND.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

components:
  schemas:
    CreateSparkRequest:
      type: object
      description: Payload for creating a new Spark.
      properties:
        title:
          type: string
          maxLength: 100
          description: Spark title (must be unique).
        description:
          type: string
          maxLength: 1500
          description: Optional description for the Spark.
      required:
        - title

    UpdateSparkRequest:
      type: object
      description: Payload for updating an existing Spark.
      properties:
        title:
          type: string
          maxLength: 100
        description:
          type: string
          maxLength: 1500
      required:
        - title

    PatchSparkRequest:
      type: object
      description: Payload for partially updating a Spark. Only present fields will be updated.
      properties:
        title:
          type: string
          maxLength: 100
          description: New title for the Spark (must be unique).
        description:
          type: string
          maxLength: 1500
          description: New description for the Spark.

    Spark:
      type: object
      description: Representation of a Spark object.
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt
        - updatedAt

    SparkTree:
      type: object
      description: Recursive representation of a Spark and its children.
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        children:
          type: array
          items:
            $ref: '#/components/schemas/SparkTree'
    SparkPage:
      type: object
      description: Paginated result containing Spark objects. `page` is zero-based.
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Spark'
        page:
          type: integer
          description: Page index (0-based).
          example: 0
        size:
          type: integer
          description: Page size.
          example: 20
        totalElements:
          type: integer
          description: Total number of matching elements.
          example: 42
        totalPages:
          type: integer
          description: Total number of pages available.
          example: 3
      required:
        - content
        - totalElements
        - totalPages
        - size
        - page

    Reason:
      type: object
      description: Representation of a Reason attached to a Spark.
      properties:
        id:
          type: string
        content:
          type: string
        type:
          $ref: '#/components/schemas/ReasonType'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - content
        - type
        - createdAt
        - updatedAt

    CreateReasonRequest:
      type: object
      description: Payload for creating a new Reason.
      properties:
        content:
          type: string
          maxLength: 1500
        type:
          $ref: '#/components/schemas/ReasonType'
      required:
        - content
        - type

    UpdateReasonRequest:
      type: object
      description: Payload for updating an existing Reason.
      properties:
        content:
          type: string
          maxLength: 1500
        type:
          $ref: '#/components/schemas/ReasonType'
      required:
        - content
        - type

    ReasonPage:
      type: object
      description: Paginated result containing Reason objects. `page` is zero-based.
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Reason'
        page:
          type: integer
          description: Page index (0-based).
          example: 0
        size:
          type: integer
          description: Page size.
          example: 10
        totalElements:
          type: integer
          description: Total number of matching elements.
          example: 2
        totalPages:
          type: integer
          description: Total number of pages available.
          example: 1
      required:
        - content
        - totalElements
        - totalPages
        - size
        - page

    ReasonType:
      type: string
      description: Reason classification.
      enum:
        - good
        - bad
      example: good

    ApiError:
      type: object
      description: Standard API error response.
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          example: 400
        error:
          type: string
          example: Bad Request
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: "title must not be blank"
        path:
          type: string
          example: /api/sparks
      required:
        - timestamp
        - status
        - error
        - code
        - message
        - path